HIVE-1806 Merge per dynamic partition based on size of each dynamic partition (Ning Zhang via namit)
git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1052414 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/ExecDriver.java
@@ -510,8 +510,7 @@ public class ExecDriver extends Task<MapredWork> implements Serializable {
       job.setNumMapTasks(work.getNumMapTasks().intValue());
     }
     if (work.getMinSplitSize() != null) {
-      HiveConf.setIntVar(job, HiveConf.ConfVars.MAPREDMINSPLITSIZE, work.getMinSplitSize()
-          .intValue());
+      HiveConf.setLongVar(job, HiveConf.ConfVars.MAPREDMINSPLITSIZE, work.getMinSplitSize().longValue());
     }
     job.setNumReduceTasks(work.getNumReduceTasks().intValue());
     job.setReducerClass(ExecReducer.class);
